feat: apply CRD at built time if `quarkus.operator-sdk.crd.apply=true`
When enabling `quarkus.operator-sdk.crd.apply=true` at build time, we will register the CRD in the Kubernetes cluster where user is logged in. 

Updated documentation and added a note saying that this will always be done when running the application on Dev mode.

Fix https://github.com/quarkiverse/quarkus-operator-sdk/issues/242
