+++ .github/workflows/build.yml
@@ -75,10 +75,9 @@ jobs:
 
           # Build manifests
           cd samples/joke
-          mvn install
+          mvn install -Dquarkus.operator-sdk.crd.apply=true
 
           # Install CRDs
-          kubectl apply -f target/kubernetes/jokerequests.samples.javaoperatorsdk.io-v1.yml
           kubectl apply -f src/main/k8s/jokes.samples.javaoperatorsdk.io-v1.yml
 
           # Run operator in DEV mode
