+++ samples/joke/README.md
@@ -14,14 +14,12 @@ cluster. You can just follow the steps below to get started quickly:
 
 - Connect to your cluster of choice using `kubectl/oc`, select the appropriate namespace/project.
   The operator will automatically connect to that cluster/namespace combination when started.
-- Run `mvn install` on the parent directory to build the project locally. This will automatically generate several resources for
-  you, in particular the CRDs associated with the custom resources we will be dealing with. These
-  CRDs are generated in `target/kubernetes/` and come in `v1` version which correspond to the versions of the CRD spec. Note that older clusters might need the version `v1beta1` which is no longer generated by default. To generate it, you need to use the property `quarkus.operator-sdk.crd.versions=v1beta1,v1`, so both versions `v1` and `v1beta1` are generated.
+- Run `mvn install -Dquarkus.operator-sdk.crd.apply=true` on the parent directory to build the project locally. 
+  This will automatically generate several resources for you, in particular the CRDs associated with the custom resources we will be dealing with, and it will also register them on the Kubernetes environment that you are logged in. These CRDs are generated in `target/kubernetes/` and come in `v1` version which correspond to the versions of the CRD spec. Note that older clusters might need the version `v1beta1` which is no longer generated by default. To generate it, you need to use the property `quarkus.operator-sdk.crd.versions=v1beta1,v1`, so both versions `v1` and `v1beta1` are generated.
 - Deploy the CRDs to your cluster (requires cluster admin privileges, while this shouldn't be an
   issue for most "testing" clusters such as `minikube` or `kind`, you might need to log in to your
   OpenShift clusters with an admin account):
   ```sh
-  kubectl apply -f target/kubernetes/jokerequests.samples.javaoperatorsdk.io-v1.yml
   kubectl apply -f src/main/k8s/jokes.samples.javaoperatorsdk.io-v1.yml
   ```           
 - Launch the app in dev mode: `mvn quarkus:dev`
